pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marco"
version = "0.1.0"
description = "CV format converter — extract structured data from CVs and render corporate templates"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "docxtpl>=0.18.0",

    "pydantic>=2.0",
    "typer>=0.12",
]

[project.scripts]
marco = "marco.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/marco"]
